3. API Test로 변경
